+++ ConfigWindow.cs
@@ -110,60 +110,9 @@ public class ConfigWindow : SimpleWindow {
 
     public override void Draw() {
         base.Draw();
-        #if !TEST
-        FestiveDecorations();
-        #endif
-        var config = SimpleTweaksPlugin.Plugin.PluginConfig;
-
 #if !TEST
-        if (config.AnalyticsOptOut == false && config.MetricsIdentifier?.Length != 64) {
-            ImGui.SetWindowFontScale(1.25f);
-            ImGui.Text("Simple Tweaks Statistics Collection");
-            ImGui.SetWindowFontScale(1f);
-            ImGui.Separator();
-            
-            ImGui.TextWrapped("" +
-                              "Simple tweaks now collects statistics of how many people have each tweak enabled. " +
-                              "This allows me (Caraxi) to get a general idea of which tweaks are actually being used and give some kind of priority to adding additional features with the same ideas. " +
-                              "By allowing collection an anonymous list of your enabled tweaks will be collected and stored on my server. \n\n" + 
-                              "You may opt out of this collection now and no information will be sent. You may also choose to opt back in at a later date in the simple tweaks config.");
-            
-            ImGui.Dummy(new Vector2(20) * ImGuiHelpers.GlobalScale);
-
-            using (ImRaii.PushColor(ImGuiCol.Button, new Vector4(0.2f, 0.7f, 0.3f, 0.8f)))
-            using (ImRaii.PushColor(ImGuiCol.ButtonActive, new Vector4(0.2f, 0.7f, 0.3f, 1f)))
-            using (ImRaii.PushColor(ImGuiCol.ButtonHovered, new Vector4(0.2f, 0.7f, 0.3f, 0.9f))) {
-                if (ImGui.Button("Allow Anonymous Statistic Collection", new Vector2(ImGui.GetContentRegionAvail().X, 40 * ImGuiHelpers.GlobalScale))) {
-                    MetricsService.ReportMetrics(true);
-                }
-            }
-
-            ImGui.Spacing();
-
-            using (ImRaii.PushColor(ImGuiCol.Button, new Vector4(0.7f, 0.2f, 0.3f, 0.8f)))
-            using (ImRaii.PushColor(ImGuiCol.ButtonActive, new Vector4(0.7f, 0.2f, 0.3f, 1f)))
-            using (ImRaii.PushColor(ImGuiCol.ButtonHovered, new Vector4(0.7f, 0.2f, 0.3f, 0.9f))) {
-                if (ImGui.Button("Disable Anonymous Statistic Collection", new Vector2(ImGui.GetContentRegionAvail().X, 25 * ImGuiHelpers.GlobalScale))) {
-                    config.AnalyticsOptOut = true;
-                    config.Save();
-                }
-            }
-
-            ImGui.Dummy(new Vector2(20) * ImGuiHelpers.GlobalScale);
-
-            ImGui.Separator();
-
-            ImGui.Dummy(new Vector2(20) * ImGuiHelpers.GlobalScale);
-
-            if (ImGui.Button("Open Changelog")) {
-                SimpleTweaksPlugin.Plugin.ChangelogWindow.IsOpen = true;
-            }
-
-            return;
-        }
-        
+        FestiveDecorations();
         SimpleTweaksPlugin.Plugin.PluginConfig.DrawConfigUI();
-        
 #else
         if (ImGui.BeginTabBar("testBar")) {
             if (ImGui.BeginTabItem("Test Runner")) {
@@ -185,6 +134,5 @@ public class ConfigWindow : SimpleWindow {
         randomDecorationType = null;
         SimpleTweaksPlugin.Plugin.SaveAllConfig();
         SimpleTweaksPlugin.Plugin.PluginConfig.ClearSearch();
-        MetricsService.ReportMetrics();
     }
 }