+++ SimpleTweaksPlugin.cs
@@ -189,8 +189,6 @@ namespace SimpleTweaksPlugin {
 
 
             Service.Framework.Update += FrameworkOnUpdate;
-            
-            MetricsService.ReportMetrics();
 
             TestUtil.Ready();
         }