+++ SimpleTweaksPluginConfig.cs
@@ -458,22 +458,6 @@ public partial class SimpleTweaksPluginConfig : IPluginConfiguration {
 
                 if (ImGui.BeginTabItem(Loc.Localize("General Options / TabHeader", "General Options") + $"###generalOptionsTab")) {
                     ImGui.BeginChild($"generalOptions-scroll", new Vector2(-1, -1));
-
-                    if (ImGui.Checkbox(Loc.Localize("General Options / Analytics Opt Out", "Opt out of metrics"), ref AnalyticsOptOut)) Save();
-                    
-                    #if DEBUG
-                    ImGui.SameLine();
-                    if (ImGui.Button("Report Metrics")) {
-                        MetricsService.ReportMetrics();
-                    }
-                    #endif
-                    
-                    
-                    ImGui.SameLine();
-                    ImGuiComponents.HelpMarker("Simple tweaks collects a list of enabled tweaks to give me an idea of which tweaks are being used. You can choose to opt out of this data collection and no information will be sent. No identifying information will be collected in any way.");
-
-                    ImGui.Separator();
-
                     if (ImGui.CollapsingHeader(Loc.Localize("General Options / Visible Category Tabs", "Visible Category Tabs") + $" ({tweakCategories.Count + (ShowAllTweaksTab ? 1 : 0) + (ShowEnabledTweaksTab ? 1 : 0)})###visibleCategoryTabs") ) {
                         ImGui.Indent();