+++ Tweaks/Tooltips/TrackOutfits.cs
@@ -24,7 +24,7 @@ namespace SimpleTweaksPlugin.Tweaks.Tooltips;
 [TweakReleaseVersion("1.10.8.0")]
 [TweakAutoConfig]
 [Changelog("1.14.0.0", "Fixed HQ outfits not working.")]
-[Changelog(UnreleasedVersion, "Added option to ignore quality when checking if an item exists as part of an outfit.")]
+[Changelog("1.15.0.5", "Added option to ignore quality when checking if an item exists as part of an outfit.")]
 public unsafe class TrackOutfits : TooltipTweaks.SubTweak
 {
     [TweakHook(typeof(UIState), nameof(UIState.IsItemActionUnlocked), nameof(IsItemActionUnlockedDetour))]