+++ Tweaks/UiAdjustment/ScenarioProgressionDisplay.cs
@@ -17,12 +17,13 @@ namespace SimpleTweaksPlugin.Tweaks.UiAdjustment;
 [TweakAutoConfig]
 [TweakReleaseVersion("1.9.0.0")]
 [Changelog("1.10.8.0", "Once again fixed logic.")]
+[Changelog("1.15.0.5", "Fixed percentage not being displayed in some circumstances.")]
 public unsafe class ScenarioProgressionDisplay : UiAdjustments.SubTweak {
-
     private ScenarioTree? finalScenario;
     private readonly Dictionary<uint, ScenarioTree> expansionBegins = new();
     private readonly Dictionary<uint, ScenarioTree> expansionEnds = new();
-
+    private SeStringBuilder? savedBuilder;
+    
     public class Config : TweakConfig {
         [TweakConfigOption("Show for current expansion", 1)]
         public bool UseCurrentExpansion;
@@ -34,7 +35,7 @@ public unsafe class ScenarioProgressionDisplay : UiAdjustments.SubTweak {
         public int Accuracy = 1;
     }
 
-    [TweakConfig] public Config TweakConfig { get; private set; }
+    [TweakConfig] public Config TweakConfig { get; private set; } = new();
 
     protected override void Enable() => UpdateAddon(Common.GetUnitBase("ScenarioTree"));
     protected override void Disable() => UpdateAddon(Common.GetUnitBase("ScenarioTree"));
@@ -52,7 +53,7 @@ public unsafe class ScenarioProgressionDisplay : UiAdjustments.SubTweak {
         if (current == null) return 0;
         
         if (finalScenario == null) {
-            foreach (var st in Service.Data.GetExcelSheet<ScenarioTree>()!) {
+            foreach (var st in Service.Data.GetExcelSheet<ScenarioTree>()) {
                 if (!Service.Data.GetExcelSheet<Quest>().TryGetRow(st.RowId, out var quest)) continue;
                 if (!quest.Expansion.IsValid) continue;
 
@@ -108,7 +109,7 @@ public unsafe class ScenarioProgressionDisplay : UiAdjustments.SubTweak {
         if (addon == null) return;
         if (addon->AtkValuesCount < 8) return;
         var textValue = addon->AtkValues + 7;
-        if (textValue->Type is not AtkValueType.String8 || textValue->String.Value == null) return;
+        if (textValue->Type is not (AtkValueType.String or AtkValueType.String8 or AtkValueType.ManagedString) || textValue->String.Value == null) return;
         var button = addon->GetComponentButtonById(13);
         if (button == null) return;
         var textNode = button->AtkComponentBase.GetTextNodeById(6);
@@ -128,7 +129,7 @@ public unsafe class ScenarioProgressionDisplay : UiAdjustments.SubTweak {
         }
         
         textNode->SetText(builder.GetViewAsSpan());
+        savedBuilder?.Clear();
+        savedBuilder = builder;
     }
-
-    public string DisabledMessage => "Tweak is not currently functional and will return when possible";
 }